+++ api/src/pdc/services/policy/engine/policies/20220412_LavalAgglo.ts
@@ -17,7 +17,7 @@ import {
   watchForPersonMaxTripByDay,
 } from '../helpers';
 import { AbstractPolicyHandler } from './AbstractPolicyHandler';
-import { description } from './Laval.html';
+import { description } from './20220412_LavalAgglo.html';
 
 // Politique de la Communauté D'Agglomeration De Laval
 export const Laval: PolicyHandlerStaticInterface = class
